move message communication.js methods to MessageService
